gitserver: Move SSH agent into separate package
To reduce the sheer size of the server package, moving this very encapsulated thing into is own package.
