+++ cmd/gitserver/server/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "server_grpc.go",
         "servermetrics.go",
         "serverutil.go",
-        "ssh_agent.go",
         "vcs_packages_syncer.go",
         "vcs_syncer.go",
         "vcs_syncer_git.go",
@@ -40,6 +39,7 @@ go_library(
         "//cmd/gitserver/server/common",
         "//cmd/gitserver/server/internal/cacert",
         "//cmd/gitserver/server/perforce",
+        "//cmd/gitserver/server/sshagent",
         "//internal/actor",
         "//internal/api",
         "//internal/codeintel/dependencies",
@@ -96,8 +96,6 @@ go_library(
         "@io_opentelemetry_go_otel//attribute",
         "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//status",
-        "@org_golang_x_crypto//ssh",
-        "@org_golang_x_crypto//ssh/agent",
         "@org_golang_x_mod//module",
         "@org_golang_x_mod//zip",
         "@org_golang_x_sync//errgroup",
@@ -116,7 +114,6 @@ go_test(
         "run_test.go",
         "server_test.go",
         "serverutil_test.go",
-        "ssh_agent_test.go",
         "vcs_packages_syncer_test.go",
         "vcs_syncer_go_modules_test.go",
         "vcs_syncer_jvm_packages_test.go",
@@ -145,7 +142,6 @@ go_test(
         "//internal/database",
         "//internal/database/dbmocks",
         "//internal/database/dbtest",
-        "//internal/encryption",
         "//internal/extsvc/gitolite",
         "//internal/extsvc/jvmpackages/coursier",
         "//internal/extsvc/npm",
@@ -176,8 +172,6 @@ go_test(
         "@com_github_stretchr_testify//require",
         "@org_golang_google_grpc//codes",
         "@org_golang_google_grpc//status",
-        "@org_golang_x_crypto//ssh",
-        "@org_golang_x_crypto//ssh/agent",
         "@org_golang_x_exp//slices",
         "@org_golang_x_mod//module",
         "@org_golang_x_sync//semaphore",
