+++ cmd/gitserver/server/sshagent/BUILD.bazel
@@ -1,0 +1,28 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
+
+go_library(
+    name = "sshagent",
+    srcs = ["ssh_agent.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/gitserver/server/sshagent",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//lib/errors",
+        "@com_github_sourcegraph_log//:log",
+        "@org_golang_x_crypto//ssh",
+        "@org_golang_x_crypto//ssh/agent",
+    ],
+)
+
+go_test(
+    name = "sshagent_test",
+    srcs = ["ssh_agent_test.go"],
+    embed = [":sshagent"],
+    deps = [
+        "//internal/encryption",
+        "//lib/errors",
+        "@com_github_sourcegraph_log//logtest",
+        "@org_golang_x_crypto//ssh",
+        "@org_golang_x_crypto//ssh/agent",
+    ],
+)
